+++ include/nc-runtime/basic_types.h
@@ -4,6 +4,7 @@
 #include <vector>
 #include <atomic>
 #include <functional>
+#include <mutex>
 #include <assert.h>
 #include <limits.h>
 #include <string.h>