+++ include/nc-runtime/nc_array.h
@@ -6,14 +6,23 @@
 template <typename T>
 class NcArray : public NcObject {
 public:
-  static sp<NcArray<T>, T> alloc() { return new NcArray(); }
+  using ArrayElement = T;
+
+public:
+  static sp<NcArray<T>> alloc() { return new NcArray(); }
 
   forceinline int capacity() { return (int)m_array.capacity(); }
   forceinline int size() { return (int)m_array.size(); }
   forceinline void reserve(int newCapacity) { return m_array.reserve(newCapacity); }
 
-  forceinline void addObject(const sp<T>& obj) { this->m_array.push_back(obj); }
+  forceinline void addObject(T* obj) { this->m_array.push_back(retain(obj)); }
+  forceinline sp<T>& firstObject() { return this->m_array[0]; }
+  forceinline sp<T>& lastObject() { return this->m_array[m_array.size() - 1]; }
   forceinline sp<T>& objectAtIndex(int i) { return this->m_array[i]; }
+  forceinline void removeObjectAtIndex(int i) {
+    assert(!m_array.empty());
+    m_array.erase(m_array.begin() + i);
+  }
 
   /**
    * Find an object
@@ -32,7 +41,12 @@ public:
    * });
    * ```
    */
-  sp<T> findWithCondition(std::function<bool (T* obj)> func);
+  template <typename Func>
+  sp<T> findWithCondition(Func& func) {
+    for (auto& obj : m_array)
+      if (func(obj.get())) return obj;
+    return NULL;
+  }
 
   /*
    * return -1 if not found. It uses NcObject::equals() for comparison
@@ -41,35 +55,15 @@ public:
     return indexOfObjectWithCondition([=](T* obj) { return obj == r || obj->equals(r); });
   }
   
-  int indexOfObjectWithCondition(std::function<bool(T* obj)> func);
-
-  forceinline sp<T>& operator[](int index) {
-    return m_array[index];
+  template<typename Func>
+  int indexOfObjectWithCondition(Func& func) {
+    int len = (int)m_array.size();
+    for (int i = 0; i < len; i++)
+      if (func(m_array[i].get()))
+        return i;
+    return -1;
   }
 
 private:
   std::vector< sp<T> > m_array;
 };
-
-template <typename T>
-sp<T> NcArray<T>::findWithCondition(std::function<bool(T* obj)> func) {
-  for (auto& obj : m_array) {
-    if (func(obj.get())) {
-      return obj;
-    }
-  }
-
-  return NULL;
-}
-
-template <typename T>
-int NcArray<T>::indexOfObjectWithCondition(std::function<bool(T* obj)> func) {
-  int len = (int)m_array.size();
-  for (int i = 0; i < len; i++) {
-    if (func(m_array[i].get())) {
-      return i;
-    }
-  }
-
-  return -1;
-}