+++ src/stdafx_nc_runtime.h
@@ -1,3 +1,6 @@
 #pragma once
 
 #include "basic_types.h"
+#include "nc_stdlib.h"
+
+using namespace std;