+++ test/nc_stdlib_unittest.cpp
@@ -23,3 +23,30 @@ TEST(Stdlib, stackOrHeapAllocator) {
 
   EXPECT_EQ(allocator._moreHeapPointerCount(), 93);
 }
+
+TEST(Stdlib, TimeTick) {
+  TimeTick start = TimeTick::now();
+  Thread::sleep(10);
+  TimeTick duration = TimeTick::now() - start;
+  EXPECT_GE(duration.ms(), 10);
+}
+
+TEST(Stdlib, ManualResetEvent) {
+  ManualResetEvent e;
+
+  std::thread t([&] {
+    Thread::sleep(100);
+    e.set();
+  });
+
+  // must wait until reset
+  EXPECT_GE(TimeTick::measure([&] { e.wait(); }).ms(), 100);
+  t.join();
+
+  // not reset(). So wait on it will not block
+  EXPECT_LT(TimeTick::measure([&] { e.wait(); }).ms(), 1);
+
+  // reset() again. Wait will block
+  e.reset();
+  EXPECT_GE(TimeTick::measure([&] { EXPECT_FALSE(e.waitWithTimeout(10)); }).ms(), 10);
+}